requirements.txt
fastapi
uvicorn
redis
kafka-python
psycopg2-binary
prometheus-client
scikit-learn
numpy
torch
stable-baselines3
joblib
